+++ controller/settingscontroller.php
@@ -127,9 +127,10 @@ class SettingsController extends Controller {
 			"compactHeader" => $this->config->getCustomizationCompactHeader(),
 			"feedback" => $this->config->getCustomizationFeedback(),
 			"forcesave" => $this->config->getCustomizationForcesave(),
+			"liveViewOnShare" => $this->config->getLiveViewOnShare(),
 			"help" => $this->config->getCustomizationHelp(),
-			"toolbarNoTabs" => $this->config->getCustomizationToolbarNoTabs(),
 			"successful" => $this->config->settingsAreSuccessful(),
+			"settingsError" => $this->config->getSettingsError(),
 			"plugins" => $this->config->getCustomizationPlugins(),
 			"macros" => $this->config->getCustomizationMacros(),
 			"reviewDisplay" => $this->config->getCustomizationReviewDisplay(),
@@ -217,8 +218,8 @@ class SettingsController extends Controller {
 	 * @param bool $compactHeader - display compact header
 	 * @param bool $feedback - display feedback
 	 * @param bool $forcesave - forcesave
+	 * @param bool $liveViewOnShare - live view on share
 	 * @param bool $help - display help
-	 * @param bool $toolbarNoTabs - display toolbar tab
 	 * @param string $reviewDisplay - review viewing mode
 	 * @param string $theme - default theme mode
 	 * @param string $unknownAuthor - display unknown author
@@ -238,8 +239,8 @@ class SettingsController extends Controller {
 		$compactHeader,
 		$feedback,
 		$forcesave,
+		$liveViewOnShare,
 		$help,
-		$toolbarNoTabs,
 		$reviewDisplay,
 		$theme,
 		$unknownAuthor
@@ -256,8 +257,8 @@ class SettingsController extends Controller {
 		$this->config->setCustomizationCompactHeader($compactHeader);
 		$this->config->setCustomizationFeedback($feedback);
 		$this->config->setCustomizationForcesave($forcesave);
+		$this->config->setLiveViewOnShare($liveViewOnShare);
 		$this->config->setCustomizationHelp($help);
-		$this->config->setCustomizationToolbarNoTabs($toolbarNoTabs);
 		$this->config->setCustomizationReviewDisplay($reviewDisplay);
 		$this->config->setCustomizationTheme($theme);
 		$this->config->setUnknownAuthor($unknownAuthor);