+++ js/main.js
@@ -604,23 +604,24 @@
           return true;
         }
 
-        const versionNodes = $("#versionsTabView ul.versions>li");
-        const versionNode = $(this).closest(
-          "#versionsTabView ul.versions>li"
-        )[0];
-
-        const href = $(this).attr("href");
-        const search = new RegExp("/meta/(\\d+)/v/\\d+");
+        const href = $("#app-sidebar .fileName .permalink").attr("href");
+        const search = new RegExp("/f/(\\d+)");
         const result = search.exec(href);
         let fileId = null;
         if (result && result.length > 1) {
           fileId = result[1];
         }
+        if (fileId) {
+          const versionNodes = $("#versionsTabView ul.versions>li");
+          const versionNode = $(this).closest(
+            "#versionsTabView ul.versions>li"
+          )[0];
 
-        const versionNum =
-          versionNodes.length - $.inArray(versionNode, versionNodes);
+          const versionNum =
+            versionNodes.length - $.inArray(versionNode, versionNodes);
 
-        OCA.Onlyoffice.openVersion(fileId || "", versionNum);
+          OCA.Onlyoffice.openVersion(fileId || "", versionNum);
+        }
 
         return false;
       }