+++ js/settings.js
@@ -68,6 +68,21 @@
     $("#onlyofficeDemo").click(demoToggle);
     demoToggle();
 
+    const connectionError = document.getElementById(
+      "onlyofficeSettingsState"
+    ).value;
+    if (connectionError !== "") {
+      const message =
+        t(OCA.Onlyoffice.AppName, "Error when trying to connect") +
+        " (" +
+        connectionError +
+        ")";
+      OC.Notification.show(message, {
+        timeout: 3,
+        type: "error",
+      });
+    }
+
     $("#onlyoffice-addr-save").click(function () {
       $(".section-onlyoffice").addClass("icon-loading");
       const onlyofficeUrl = $("#onlyofficeUrl").val().trim();
@@ -79,18 +94,16 @@
       }
 
       const onlyofficeInternalUrl = (
-        $("#onlyofficeInternalUrl:visible").val() || ""
+        $("#onlyofficeInternalUrl").val() || ""
       ).trim();
       const onlyofficeStorageUrl = (
-        $("#onlyofficeStorageUrl:visible").val() || ""
+        $("#onlyofficeStorageUrl").val() || ""
       ).trim();
       const onlyofficeVerifyPeerOff = $("#onlyofficeVerifyPeerOff").prop(
         "checked"
       );
-      const onlyofficeSecret = (
-        $("#onlyofficeSecret:visible").val() || ""
-      ).trim();
-      const jwtHeader = ($("#onlyofficeJwtHeader:visible").val() || "").trim();
+      const onlyofficeSecret = ($("#onlyofficeSecret").val() || "").trim();
+      const jwtHeader = ($("#onlyofficeJwtHeader").val() || "").trim();
       const demo = $("#onlyofficeDemo").prop("checked");
 
       $.ajax({
@@ -167,7 +180,9 @@
       const sameTab = $("#onlyofficeSameTab").is(":checked");
       const preview = $("#onlyofficePreview").is(":checked");
       const cronChecker = $("#onlyofficeCronChecker").is(":checked");
-      const emailNotifications = $('#onlyofficeEmailNotifications').is(':checked');
+      const emailNotifications = $("#onlyofficeEmailNotifications").is(
+        ":checked"
+      );
       const versionHistory = $("#onlyofficeVersionHistory").is(":checked");
 
       const limitGroupsString = $("#onlyofficeGroups").prop("checked")
@@ -179,8 +194,8 @@
       const compactHeader = $("#onlyofficeCompactHeader").is(":checked");
       const feedback = $("#onlyofficeFeedback").is(":checked");
       const forcesave = $("#onlyofficeForcesave").is(":checked");
+      const liveViewOnShare = $("#onlyofficeLiveViewOnShare").is(":checked");
       const help = $("#onlyofficeHelp").is(":checked");
-      const toolbarNoTabs = $("#onlyofficeToolbarNoTabs").is(":checked");
       const reviewDisplay = $(
         "input[type='radio'][name='reviewDisplay']:checked"
       )
@@ -189,7 +204,7 @@
       const theme = $("input[type='radio'][name='theme']:checked")
         .attr("id")
         .replace("onlyofficeTheme_", "");
-      const unknownAuthor = $('#onlyofficeUnknownAuthor').val().trim();
+      const unknownAuthor = $("#onlyofficeUnknownAuthor").val().trim();
 
       $.ajax({
         method: "PUT",
@@ -209,8 +224,8 @@
           compactHeader,
           feedback,
           forcesave,
+          liveViewOnShare,
           help,
-          toolbarNoTabs,
           reviewDisplay,
           theme,
           unknownAuthor,