+++ l10n/pt_BR.js
@@ -58,7 +58,6 @@ OC.L10N.register(
     "Display the header more compact" : "Exibir o cabeçalho mais compacto",
     "Display Feedback & Support menu button" : "Exibir o botão do menu de feedback e suporte",
     "Display Help menu button" : "Exibir o botão do menu Ajuda",
-    "Display monochrome toolbar header" : "Exibir cabeçalho da barra de ferramentas monocromático",
     "Save as" : "Salvar como",
     "File saved" : "O arquivo foi salvo",
     "Insert image" : "Inserir imagem",
@@ -107,7 +106,6 @@ OC.L10N.register(
     "Run document macros": "Executar macros de documento",
     "Default editor theme": "Tema do editor padrão",
     "Light": "Claro",
-    "Classic Light": "Clássico claro",
     "Dark": "Escuro",
     "This feature is unavailable due to encryption settings.": "Este recurso não está disponível devido às configurações de criptografia.",
     "Enable plugins": "Ativar plug-ins",
@@ -138,7 +136,6 @@ OC.L10N.register(
     "Select file to insert text": "Selecionar arquivo para inserir texto",
     "Select file": "Selecionar arquivo",
     "Same as system": "O mesmo que sistema",
-    "Contrast dark": "Contraste escuro",
-    "Gray": "Cinza"
+    "Enable live-viewing mode when accessing file by public link" : "Habilitar o modo de visualização ao vivo ao acessar o arquivo por link público"
 },
 "nplurals=2; plural=(n > 1);");