+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 9.10.1
+## Added
+- insert svg to editor
+- view vsdx, vssx, vstx, vsdm, vssm, vstm
+- view odg, md
+- setting for enabling live-view for shared docs
+- refresh file when connection is restored
+
+## Changed
+- edit xlsb
+- customization_goback setting removed
+- there are only system, light and dark themes
+- toolbarNoTabs settings removed
+- fix opening version from sidebar
+- display error when opening settings page if it exists
+
 ## 9.9.1
 ## Added
 - support external link for reference data